+++ src/containers/Orders/UploadStandardTemplateModal/index.js
@@ -5,6 +5,7 @@ import { connect } from 'react-redux';
 import { createStructuredSelector } from 'reselect';
 import { compose } from 'redux';
 import * as excelDataTransfer from 'utils/excelDataTransfer';
+import { notification } from 'antd';
 
 import {
   makeSelectMAWB,
@@ -52,59 +53,66 @@ import {
 
 function UploadStandardTemplateModal(props) {
   async function readFile() {
-    let isDetailInfoSheet;
-    const input = document.getElementById('xlsxStandardInput');
-    input.addEventListener('change', () => {
-      if (input.files.length !== 0) {
-        const dataList = [];
-        props.setUploadingPercent(0);
-        props.setFiles(input.files);
-        readXlsxFile(input.files[0], { getSheets: true }).then(sheets => {
-          isDetailInfoSheet = sheets.find(sheet => sheet === 'Detail info') !== undefined ? true : false;
-        });
-        readXlsxFile(input.files[0], { sheet: 'Detail info' }).then(async rows => {
-          for (let index = 0; index < rows.length; index++) {
-            props.setUploadingPercent(Math.ceil(index / rows.length) * 100);
+    try {
+      let isDetailInfoSheet;
+      const input = document.getElementById('xlsxStandardInput');
+      input.addEventListener('change', () => {
+        if (input.files.length !== 0) {
+          const dataList = [];
+          props.setUploadingPercent(0);
+          props.setFiles(input.files);
+          readXlsxFile(input.files[0], { getSheets: true }).then(sheets => {
+            isDetailInfoSheet = sheets.find(sheet => sheet === 'Detail info') !== undefined ? true : false;
+          });
+          readXlsxFile(input.files[0], { sheet: 'Detail info' }).then(async rows => {
+            for (let index = 0; index < rows.length; index++) {
+              props.setUploadingPercent(Math.ceil(index / rows.length) * 100);
 
-            if (index === 1) {
-              if (rows[index].length !== 19) {
-                return;
+              if (index === 1) {
+                if (rows[index].length !== 19) {
+                  return;
+                }
+              }
+              if (index >= 2) {
+                const row = rows[index];
+                const data = {
+                  MAWB: excelDataTransfer.stringTrans(row[0]),
+                  containerNumber: excelDataTransfer.stringTrans(row[1]),
+                  trackingNumber: excelDataTransfer.stringTrans(row[2]),
+                  shipper: excelDataTransfer.stringTrans(row[3]),
+                  shipperPhoneNumber: excelDataTransfer.stringTrans(row[4]),
+                  shipperAddress: excelDataTransfer.stringTrans(row[5]),
+                  destinationCountry: excelDataTransfer.stringTrans(row[6]),
+                  recipient: excelDataTransfer.stringTrans(row[7]),
+                  RUT: excelDataTransfer.stringTrans(row[8]),
+                  recipientPhoneNumber: excelDataTransfer.stringTrans(row[9]),
+                  recipientEmail: excelDataTransfer.stringTrans(row[10]),
+                  region: excelDataTransfer.stringTrans(row[11]),
+                  province: excelDataTransfer.stringTrans(row[12]),
+                  comuna: excelDataTransfer.stringTrans(row[13]),
+                  address: excelDataTransfer.stringTrans(row[14]),
+                  weight: excelDataTransfer.numberTrans(row[15]),
+                  value: excelDataTransfer.numberTrans(row[16]),
+                  description: excelDataTransfer.stringTrans(row[17]),
+                  quantity: excelDataTransfer.numberTrans(row[18]),
+                };
+                dataList.push(data);
               }
             }
-            if (index >= 2) {
-              const row = rows[index];
-              const data = {
-                MAWB: excelDataTransfer.stringTrans(row[0]),
-                containerNumber: excelDataTransfer.stringTrans(row[1]),
-                trackingNumber: excelDataTransfer.stringTrans(row[2]),
-                shipper: excelDataTransfer.stringTrans(row[3]),
-                shipperPhoneNumber: excelDataTransfer.stringTrans(row[4]),
-                shipperAddress: excelDataTransfer.stringTrans(row[5]),
-                destinationCountry: excelDataTransfer.stringTrans(row[6]),
-                recipient: excelDataTransfer.stringTrans(row[7]),
-                RUT: excelDataTransfer.stringTrans(row[8]),
-                recipientPhoneNumber: excelDataTransfer.stringTrans(row[9]),
-                recipientEmail: excelDataTransfer.stringTrans(row[10]),
-                region: excelDataTransfer.stringTrans(row[11]),
-                province: excelDataTransfer.stringTrans(row[12]),
-                comuna: excelDataTransfer.stringTrans(row[13]),
-                address: excelDataTransfer.stringTrans(row[14]),
-                weight: excelDataTransfer.numberTrans(row[15]),
-                value: excelDataTransfer.numberTrans(row[16]),
-                description: excelDataTransfer.stringTrans(row[17]),
-                quantity: excelDataTransfer.numberTrans(row[18]),
-              };
-              dataList.push(data);
-            }
-          }
-          props.setUploadingDataList(dataList);
-        });
-      } else {
-        props.setUploadingPercent(100);
-        props.setFiles([]);
-        props.setUploadingDataList([]);
-      }
-    });
+            props.setUploadingDataList(dataList);
+          });
+        } else {
+          props.setUploadingPercent(100);
+          props.setFiles([]);
+          props.setUploadingDataList([]);
+        }
+      });
+    } catch (err) {
+      console.error(err);
+      notification['error']({
+        message: 'Format Error, please check uploaded file',
+      });
+    }
   }
 
   React.useEffect(() => {